Implementing a general model for the first transformer
Trying to allow full song input.

Note: Not finished, just uploading progress.
