--- resources/logs/used_songs_one_token.log
@@ -1,75 +1,0 @@
-Song 0 - 2 were used.
-Song 2 - 4 were used.
-Song 4 - 6 were used.
-Song 6 - 8 were used.
-Song 8 - 10 were used.
-Song 10 - 12 were used.
-Song 12 - 14 were used.
-Song 14 - 16 were used.
-Song 16 - 18 were used.
-Song 18 - 20 were used.
-Song 20 - 22 were used.
-Song 22 - 24 were used.
-Song 24 - 26 were used.
-Song 26 - 28 were used.
-Song 28 - 30 were used.
-Song 30 - 32 were used.
-Song 32 - 34 were used.
-Song 34 - 36 were used.
-Song 36 - 38 were used.
-Song 38 - 40 were used.
-Song 40 - 42 were used.
-Song 42 - 44 were used.
-Song 44 - 46 were used.
-Song 46 - 48 were used.
-Song 48 - 50 were used.
-Song 50 - 52 were used.
-Song 52 - 54 were used.
-Song 54 - 56 were used.
-Song 56 - 58 were used.
-Song 58 - 60 were used.
-Song 60 - 62 were used.
-Song 62 - 64 were used.
-Song 64 - 66 were used.
-Song 66 - 68 were used.
-Song 68 - 70 were used.
-Song 70 - 72 were used.
-Song 72 - 74 were used.
-Song 74 - 76 were used.
-Song 76 - 78 were used.
-Song 78 - 80 were used.
-Song 80 - 82 were used.
-Song 82 - 84 were used.
-Song 84 - 86 were used.
-Song 86 - 88 were used.
-Song 88 - 90 were used.
-Song 90 - 92 were used.
-Song 92 - 94 were used.
-Song 94 - 96 were used.
-Song 96 - 98 were used.
-Song 98 - 100 were used.
-Song 100 - 104 were used.
-Song 104 - 108 were used.
-Song 108 - 112 were used.
-Song 112 - 116 were used.
-Song 116 - 120 were used.
-Song 120 - 124 were used.
-Song 124 - 128 were used.
-Song 128 - 132 were used.
-Song 132 - 136 were used.
-Song 136 - 140 were used.
-Song 140 - 144 were used.
-Song 144 - 148 were used.
-Song 148 - 152 were used.
-Song 152 - 156 were used.
-Song 156 - 160 were used.
-Song 160 - 164 were used.
-Song 164 - 168 were used.
-Song 168 - 172 were used.
-Song 172 - 176 were used.
-Song 176 - 180 were used.
-Song 180 - 184 were used.
-Song 184 - 188 were used.
-Song 188 - 192 were used.
-Song 192 - 196 were used.
-Song 196 - 200 were used.
